get.sh with questions

+++ software/get.sh
@@ -17,6 +17,15 @@ set -e
 #   $ sudo bash get.sh
 #
 
+##################################################
+
+echo "What shall be installed?"
+read -p "Register shelf-controller? (y/n): " response01
+read -p "Register TOF-controller? (y/n): " response02
+read -p "Register display power control? (y/n): " response03
+read -p "Install as a basket display? (y/n): " response04
+
+##################################################
 
 sstr="**********************************************************************\n"
 
@@ -42,12 +51,21 @@ sudo -u pi pip3 install --break-system-packages -r requirements.txt
 
 #################################################
 echo -e "$sstr   Register zeitlos software in systemd \n$sstr"
-echo -e "$sstr   Register shelf-controller \n$sstr"
-./zeitlos/software/shelf-controller/install.sh
-echo -e "$sstr   Register TOF-controller \n$sstr"
-./zeitlos/software/waveshare-tof-readout/install.sh
-echo -e "$sstr   Register display power control \n$sstr"
-./zeitlos/software/display-power-control/install.sh
+
+if [[ "$response01" == "y" ]]; then
+    echo -e "$sstr   Register shelf-controller \n$sstr"
+    ./zeitlos/software/shelf-controller/install.sh
+fi
+
+if [[ "$response02" == "y" ]]; then
+    echo -e "$sstr   Register TOF-controller \n$sstr"
+    ./zeitlos/software/waveshare-tof-readout/install.sh
+fi
+
+if [[ "$response03" == "y" ]]; then
+    echo -e "$sstr   Register display power control \n$sstr"
+    ./zeitlos/software/display-power-control/install.sh
+fi
 
 echo -e "$sstr   ALL zeitlos SOFTWARE INSTALLED. \n$sstr"
 
@@ -70,57 +88,61 @@ apt-get purge -y dphys-swapfile
 #enable ssh
 # 0 - Enable SSH
 # 1 - Disable SSH
+echo "enable SSH login"
 raspi-config nonint do_ssh 0
 #
 #Select a locale, for example en_GB.UTF-8 UTF-8.
 #correct setting will display currency and time correctly
+echo "change to de locale"
 raspi-config nonint do_change_locale de_DE.UTF-8
 #
 #timezone
+echo "set timezone"
 raspi-config nonint do_change_timezone Europe/Berlin
-
-
-###################################
-# for shelf displays
-# 
-# necessary to have the standard or full versino of Raspberry OS 
-# installed. Will not work with the lite version
-###################################
-
-###################################
-# get rid of the welcome wizard
-# see: https://forums.raspberrypi.com/viewtopic.php?t=231557
-###################################
-echo -e "$sstr   Delete file piwiz.desktop \n$sstr"
-myfile="/etc/xdg/autostart/piwiz.desktop"
-[ -e $myfile ] && rm $myfile && echo "File piwiz.desktop deleted."
-###################################
-
-
-echo -e "$sstr   set boot target do_boot_behaviour \n$sstr"
-#boot to (B1=console, requiring login)
-#        (B4 - Boot to desktop, logging in automatically)
-raspi-config nonint do_boot_behaviour B4
 
 # 0 - Enable splash screen
 # 1 - Disable splash screen
 echo -e "$sstr   raspi-config nonint do_boot_splash 1 \n$sstr"
 raspi-config nonint do_boot_splash 1
 
-##screen blanking, default=1 (=Yes)
-# details here: https://www.raspberrypi.com/documentation/computers/configuration.html#desktop
-echo -e "$sstr   raspi-config nonint do_blanking 0 \n$sstr"
-raspi-config nonint do_blanking 0 
+###################################
+# for shelf displays
+#
+# necessary to have the standard or full versino of Raspberry OS
+# installed. Will not work with the lite version
+###################################
 
-##activate VNC, default=1 (0=enabled, 1=No)
-echo -e "$sstr   raspi-config nonint do_vnc 0  \n$sstr"
-raspi-config nonint do_vnc 0 
 
-##browser kiosk mode
-echo -e "$sstr   activate browser kiosk mode \n$sstr"
-sudo apt install wtype
+if [[ "$response04" == "y" ]]; then
+    ###################################
+    # get rid of the welcome wizard
+    # see: https://forums.raspberrypi.com/viewtopic.php?t=231557
+    ###################################
+    echo -e "$sstr   Delete file piwiz.desktop \n$sstr"
+    myfile="/etc/xdg/autostart/piwiz.desktop"
+    [ -e $myfile ] && rm $myfile && echo "File piwiz.desktop deleted."
+    ###################################
 
-cat <<EOT >> /home/pi/.config/wayfire.ini
+
+    echo -e "$sstr   set boot target do_boot_behaviour \n$sstr"
+    #boot to (B1=console, requiring login)
+    #        (B4 - Boot to desktop, logging in automatically)
+    raspi-config nonint do_boot_behaviour B4
+
+    ##screen blanking, default=1 (=Yes)
+    # details here: https://www.raspberrypi.com/documentation/computers/configuration.html#desktop
+    echo -e "$sstr   raspi-config nonint do_blanking 0 \n$sstr"
+    raspi-config nonint do_blanking 0 
+
+    ##activate VNC, default=1 (0=enabled, 1=No)
+    echo -e "$sstr   raspi-config nonint do_vnc 0  \n$sstr"
+    raspi-config nonint do_vnc 0 
+
+    ##browser kiosk mode
+    echo -e "$sstr   activate browser kiosk mode \n$sstr"
+    sudo apt install wtype
+
+    cat <<EOT >> /home/pi/.config/wayfire.ini
 [autostart]
 panel = wfrespawn wf-panel-pi
 background = wfrespawn pcmanfm --desktop --profile LXDE-pi
@@ -132,6 +154,12 @@ EOT
 
 #########################################
 
+else
+    echo -e "$sstr   set boot target do_boot_behaviour \n$sstr"
+    #boot to (B1=console, requiring login)
+    #        (B4 - Boot to desktop, logging in automatically)
+    raspi-config nonint do_boot_behaviour B1
+fi
 
 #Overlay file system, 0=enabled, 1 =disable
 #raspi-config nonint do_overlayfs 0
